fix: batch PUT updates, timing attack in auth, refs type, compression test
   - Use update_key_atomic in batch PUT to allow key updates (was failing)
   - Fix constant-time comparison to not leak token length
   - Change refs field from u32 to u64 for client API consistency
   - Fix compression test to use valid level 9 (was testing 10, clamped to 9)

+++ src/server/handlers/list.rs
@@ -12,7 +12,7 @@ struct KeyInfo {
     size: u64,
     hash: String,
     hash_algorithm: String,
-    refs: u32,
+    refs: u64,
     created_at: u64,
 }
